enhanced_requirements.txt
# Core Django Framework
Django==3.1.12
djongo==1.3.7
pymongo==3.11.4
djangorestframework==3.11.2
django-cors-headers==3.10.1

# Face Recognition and Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
mediapipe>=0.10.0
deepface>=0.0.79
tensorflow>=2.13.0
scikit-learn>=1.3.0

# Image Processing and Analysis
dlib>=19.24.0
face-recognition>=1.3.0
mtcnn>=0.1.1
facenet-pytorch>=2.5.3

# Machine Learning and AI
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Data Processing and Analysis
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Security and Encryption
cryptography>=41.0.0
pycryptodome>=3.18.0
bcrypt>=4.0.0

# Web and API
requests>=2.31.0
urllib3>=2.0.0
httpx>=0.24.0

# Configuration and Environment
python-decouple>=3.8
python-dotenv>=1.0.0

# Database and Caching
redis>=4.6.0
celery>=5.3.0

# Monitoring and Logging
sentry-sdk>=1.30.0
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Development Tools
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Production Deployment
gunicorn>=21.2.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.0

# Additional Utilities
python-dateutil>=2.8.0
pytz>=2023.3
jsonschema>=4.19.0
